Move simd library closer to hash join
