+++ ydb/library/yql/dq/comp_nodes/hash_join_utils/hashes_calc.h
@@ -2,7 +2,7 @@
 
 #include <util/system/unaligned_mem.h>
 
-#include <ydb/library/yql/utils/simd/simd.h>
+#include <ydb/library/yql/dq/comp_nodes/hash_join_utils/simd/simd.h>
 
 namespace NKikimr {
 namespace NMiniKQL {
