+++ ydb/library/yql/dq/comp_nodes/hash_join_utils/simd/exec/add_columns/ya.make
@@ -6,6 +6,6 @@ SIZE(MEDIUM)
 
 CFLAGS(-mavx2)
 
-PEERDIR(ydb/library/yql/utils/simd)
+PEERDIR(ydb/library/yql/dq/comp_nodes/hash_join_utils/simd)
 
 END()
