+++ ydb/library/yql/dq/comp_nodes/hash_join_utils/simd/exec/pack_tuple/ya.make
@@ -1,0 +1,9 @@
+PROGRAM()
+
+SRCS(main.cpp)
+
+CFLAGS(-mavx2)
+
+PEERDIR(ydb/library/yql/dq/comp_nodes/hash_join_utils/simd)
+
+END()
