+++ ydb/library/yql/dq/comp_nodes/hash_join_utils/simd/exec/runtime_dispatching/SSE42_algo/sse42_algo.cpp
@@ -1,4 +1,4 @@
-#include <ydb/library/yql/utils/simd/exec/runtime_dispatching/algo.h>
+#include <ydb/library/yql/dq/comp_nodes/hash_join_utils/simd/exec/runtime_dispatching/algo.h>
 #include <immintrin.h>
 
 template<>
