+++ ydb/library/yql/dq/comp_nodes/hash_join_utils/simd/exec/runtime_dispatching/main/main.cpp
@@ -1,4 +1,4 @@
-#include <ydb/library/yql/utils/simd/exec/runtime_dispatching/algo.h>
+#include <ydb/library/yql/dq/comp_nodes/hash_join_utils/simd/exec/runtime_dispatching/algo.h>
 
 int main() {
