+++ ydb/library/yql/dq/comp_nodes/hash_join_utils/simd/exec/tuples_to_bucket/main.cpp
@@ -5,7 +5,7 @@
 
 #include <cstdlib>
 
-#include <ydb/library/yql/utils/simd/simd.h>
+#include <ydb/library/yql/dq/comp_nodes/hash_join_utils/simd/simd.h>
 
 struct TPerfomancer {
     TPerfomancer() = default;
