+++ ydb/library/yql/dq/comp_nodes/hash_join_utils/simd/exec/tuples_to_bucket/ya.make
@@ -1,0 +1,7 @@
+PROGRAM()
+
+SRCS(main.cpp)
+
+PEERDIR(ydb/library/yql/dq/comp_nodes/hash_join_utils/simd)
+
+END()
