+++ ydb/library/yql/dq/comp_nodes/hash_join_utils/simd/ut/ya.make
@@ -1,0 +1,7 @@
+UNITTEST_FOR(ydb/library/yql/dq/comp_nodes/hash_join_utils/simd)
+
+SRCS(
+    simd_ut.cpp
+)
+
+END()
