+++ ydb/library/yql/dq/comp_nodes/hash_join_utils/tuple.h
@@ -6,7 +6,7 @@
 
 #include <util/generic/buffer.h>
 
-#include <ydb/library/yql/utils/simd/simd.h>
+#include <ydb/library/yql/dq/comp_nodes/hash_join_utils/simd/simd.h>
 
 namespace NKikimr {
 namespace NMiniKQL {
