+++ ydb/library/yql/dq/comp_nodes/hash_join_utils/ya.make
@@ -7,6 +7,7 @@ PEERDIR(
     yql/essentials/utils
     yql/essentials/utils/log
     library/cpp/digest/crc32c
+    ydb/library/yql/dq/comp_nodes/hash_join_utils/simd
 )
 
 IF (ARCH_X86_64)
@@ -25,6 +26,10 @@ ENDIF()
 YQL_LAST_ABI_VERSION()
 
 END()
+
+RECURSE(
+    simd
+)
 
 RECURSE_FOR_TESTS(
     ut
