--- ydb/library/yql/utils/simd/exec/runtime_dispatching/main/ya.make
@@ -1,7 +1,0 @@
-PROGRAM()
-
-SRCS(main.cpp)
-
-PEERDIR(ydb/library/yql/utils/simd/exec/runtime_dispatching)
-
-END()
