+++ ydb/library/yql/utils/ya.make
@@ -4,6 +4,5 @@ RECURSE(
     actor_system
     bindings
     plan
-    simd
 )
